wip: defining structs and fucntions in the new structure
