+++ src/services/task_placer.rs
